+++ tests/bootstrap.php
@@ -1,0 +1,5 @@
+<?php
+
+$loader = require __DIR__.'/../vendor/autoload.php';
+$loader->add('Neutron\Tests', __DIR__);
+$loader->add('Neutron\Functional', __DIR__);