+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 CHANGELOG
 =========
 
+* 0.2.0 (2013-xx-xx)
+
+  * `Clock` is now an event emitter that emits 'tick' events.
+  * BC Break : `Clock::addTimer` and `Clock::addPeriodicTimer` now returns
+    `Timer` object (thanks to react implementation).
+  * New Timer API.
+  * Support for multiple clocks in the same PHP process.
+  * New methods `Clock::pause` and `Clock::resume`.
+  * Implemented proper destruction method `Clock::destroy`.
+
 * 0.1.2 (2013-05-13)
 
   * Pass signature to timers callbacks.